requirements.txt
django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
django-cors-headers==3.12.0